ZIUM_Control/requirements.txt
# pick2build pip dependencies not available via rosdep
# Install with: pip install -r requirements.txt

mediapipe          # detection.py — hand landmark detection (MediaPipe)
ultralytics        # yolo.py — YOLO v8 object detection
sounddevice        # stt.py — audio recording
pyaudio            # MicController.py, get_keyword.py — microphone stream (PyAudio)
openai             # stt.py — Whisper STT API
langchain-openai   # get_keyword.py — ChatOpenAI (GPT-4o)
langchain          # get_keyword.py — PromptTemplate
python-dotenv      # get_keyword.py — .env loading
